Fixed SMP errors
